% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Definitions.R
\docType{data}
\name{TableDef}
\alias{TableDef}
\title{TableDef}
\format{
An object of class \code{proto} (inherits from \code{environment}) of length 26.
}
\usage{
TableDef
}
\description{
Prototype object that holds the defining factors
of a Data Table. This object is of class "def" including class
"proto" and "environment". Use this class to define the constraints
and characteristics of a Data Table.
}
\section{Feilds}{


The fields specify attributes of a Data Table. Not all fields are
necessary, but those in bold should be defined. The associated set
method is displayed in which it takes one argument, its associated
field. Most set methods are just convienice and help standardize
field expectations. Not all fields have a set method as they are
updated by another Method. Most fields have an assocated valid{Field}
method that simply checks if the associated field is NA or not.

\itemize{
\item{data}{ \code{setTableData(data)}.
holds the Data Table in memory. This field is not necessary,
but is a convience for programming.setMethod: setTableData}
\item{\strong{name}}{ \code{setTableName(name)}.
holds the character string that defines the name of the Data
Table. Each Data Table name should be unique within a Data Manager.}
\item{\strong{file}}{ \code{setTableFile(file)}. holds the full path to a file where the data will be stored
on disk.}
\item{\strong{col_names}}{ \code{setTableColnames(col_names)}. Specifies the Column names of the Data Table.}
\item{\strong{col_types}}{ \code{setTableColtypes(col_types)}. Specifies the Column classes of the Data Table.}
\item{keys}{ \code{setTableKeys(keys)}. Specifies which Columns are keys. Keys are not necessary,
but a convience. In some Data Table designs, a set of keys may express
a set of columns that uniquely define each row within a Data Table
and/or express which keys may be joined with other Data Tables in
a Data Manager.}
\item{md5sum}{ Stores the md5sum that is stored in the file field. md5sum should not be set, instead it is updated with \code{updateClass()} method.}
}


Methods to set each field
}

\keyword{datasets}
