% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Definitions.R
\name{TableDef}
\alias{TableDef}
\alias{DataManR}
\title{TableDef}
\description{
Prototype object that holds the defining factors
of a Data Table. This object is of class "def" including class
"proto" and "environment". Use this class to define the constraints
and characteristics of a Data Table.

Methods of TableDef
}
\details{
Methods to set each field
}
\section{Feilds}{


The fields specify attributes of a Data Table. Not all fields are
necessary, but those in bold should be defined. The associated set
method is displayed in which it takes one argument, its associated
field. Most set methods are just convienice and help standardize
field expectations. Not all fields have a set method as they are
updated by another Method. Most fields have an assocated valid{Field}
method that simply checks if the associated field is NA or not.

\itemize{
\item{data}{ \code{setTableData(data)}.
holds the Data Table in memory. This field is not necessary,
but is a convience for programming.setMethod: setTableData}
\item{\strong{name}}{ \code{setTableName(name)}.
holds the character string that defines the name of the Data
Table. Each Data Table name should be unique within a Data Manager.}
\item{\strong{file}}{ \code{setTableFile(file)}. holds the full path to a file where the data will be stored
on disk.}
\item{\strong{col_names}}{ \code{setTableColnames(col_names)}. Specifies the Column names of the Data Table.}
\item{\strong{col_types}}{ \code{setTableColtypes(col_types)}. Specifies the Column classes of the Data Table.}
\item{keys}{ \code{setTableKeys(keys)}. Specifies which Columns are keys. Keys are not necessary,
but a convience. In some Data Table designs, a set of keys may express
a set of columns that uniquely define each row within a Data Table
and/or express which keys may be joined with other Data Tables in
a Data Manager.}
\item{md5sum}{ Stores the md5sum that is stored in the file field. md5sum should not be set, instead it is updated with \code{updateClass()} method.}
}
}

\section{Super class}{
\code{\link[datamanR:BaseDMR]{datamanR::BaseDMR}} -> \code{DataManR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-show_link}{\code{DataManR$show_link()}}
\item \href{#method-setTableLink}{\code{DataManR$setTableLink()}}
\item \href{#method-pull_key_pairs}{\code{DataManR$pull_key_pairs()}}
\item \href{#method-full_join}{\code{DataManR$full_join()}}
\item \href{#method-addTable}{\code{DataManR$addTable()}}
\item \href{#method-rmTable}{\code{DataManR$rmTable()}}
\item \href{#method-save}{\code{DataManR$save()}}
\item \href{#method-copy}{\code{DataManR$copy()}}
\item \href{#method-new}{\code{DataManR$new()}}
\item \href{#method-validate}{\code{DataManR$validate()}}
\item \href{#method-isValid}{\code{DataManR$isValid()}}
\item \href{#method-print}{\code{DataManR$print()}}
\item \href{#method-clone}{\code{DataManR$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="datamanR" data-topic="BaseDMR" data-id="history">}\href{../../datamanR/html/BaseDMR.html#method-history}{\code{datamanR::BaseDMR$history()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="datamanR" data-topic="BaseDMR" data-id="push_history">}\href{../../datamanR/html/BaseDMR.html#method-push_history}{\code{datamanR::BaseDMR$push_history()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_link"></a>}}
\if{latex}{\out{\hypertarget{method-show_link}{}}}
\subsection{Method \code{show_link()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$show_link()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTableLink"></a>}}
\if{latex}{\out{\hypertarget{method-setTableLink}{}}}
\subsection{Method \code{setTableLink()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$setTableLink(LeftTable, LeftKey, RightTable, RightKey)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_key_pairs"></a>}}
\if{latex}{\out{\hypertarget{method-pull_key_pairs}{}}}
\subsection{Method \code{pull_key_pairs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$pull_key_pairs(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_join"></a>}}
\if{latex}{\out{\hypertarget{method-full_join}{}}}
\subsection{Method \code{full_join()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$full_join(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTable"></a>}}
\if{latex}{\out{\hypertarget{method-addTable}{}}}
\subsection{Method \code{addTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$addTable(def)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rmTable"></a>}}
\if{latex}{\out{\hypertarget{method-rmTable}{}}}
\subsection{Method \code{rmTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$rmTable(def)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$save(file = self$rds_file, saveData = F)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{file to write. Defaults to file field.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\if{latex}{\out{\hypertarget{method-copy}{}}}
\subsection{Method \code{copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$copy(deep = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$new(name = NA_character_, path = NA_character_, Tables = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-isValid}{}}}
\subsection{Method \code{isValid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$isValid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
