% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Definitions.R
\name{TableDef_methods}
\alias{TableDef_methods}
\alias{DataManR}
\title{TableDef Methods}
\description{
Methods of TableDef
}
\section{Super class}{
\code{\link[datamanR:BaseDMR]{datamanR::BaseDMR}} -> \code{DataManR}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-show_link}{\code{DataManR$show_link()}}
\item \href{#method-setTableLink}{\code{DataManR$setTableLink()}}
\item \href{#method-pull_key_pairs}{\code{DataManR$pull_key_pairs()}}
\item \href{#method-full_join}{\code{DataManR$full_join()}}
\item \href{#method-addTable}{\code{DataManR$addTable()}}
\item \href{#method-rmTable}{\code{DataManR$rmTable()}}
\item \href{#method-save}{\code{DataManR$save()}}
\item \href{#method-copy}{\code{DataManR$copy()}}
\item \href{#method-new}{\code{DataManR$new()}}
\item \href{#method-validate}{\code{DataManR$validate()}}
\item \href{#method-isValid}{\code{DataManR$isValid()}}
\item \href{#method-print}{\code{DataManR$print()}}
\item \href{#method-clone}{\code{DataManR$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="datamanR" data-topic="BaseDMR" data-id="history">}\href{../../datamanR/html/BaseDMR.html#method-history}{\code{datamanR::BaseDMR$history()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="datamanR" data-topic="BaseDMR" data-id="push_history">}\href{../../datamanR/html/BaseDMR.html#method-push_history}{\code{datamanR::BaseDMR$push_history()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show_link"></a>}}
\if{latex}{\out{\hypertarget{method-show_link}{}}}
\subsection{Method \code{show_link()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$show_link()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setTableLink"></a>}}
\if{latex}{\out{\hypertarget{method-setTableLink}{}}}
\subsection{Method \code{setTableLink()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$setTableLink(LeftTable, LeftKey, RightTable, RightKey)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pull_key_pairs"></a>}}
\if{latex}{\out{\hypertarget{method-pull_key_pairs}{}}}
\subsection{Method \code{pull_key_pairs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$pull_key_pairs(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-full_join"></a>}}
\if{latex}{\out{\hypertarget{method-full_join}{}}}
\subsection{Method \code{full_join()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$full_join(x, y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-addTable"></a>}}
\if{latex}{\out{\hypertarget{method-addTable}{}}}
\subsection{Method \code{addTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$addTable(def)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rmTable"></a>}}
\if{latex}{\out{\hypertarget{method-rmTable}{}}}
\subsection{Method \code{rmTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$rmTable(def)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$save(
  location = dirname(self$path),
  name = str_c(self$name, "_DataManR.rds"),
  saveData = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\if{latex}{\out{\hypertarget{method-copy}{}}}
\subsection{Method \code{copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$copy(deep = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$new(name = NA_character_, path = NA_character_, Tables = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate"></a>}}
\if{latex}{\out{\hypertarget{method-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-isValid"></a>}}
\if{latex}{\out{\hypertarget{method-isValid}{}}}
\subsection{Method \code{isValid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$isValid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DataManR$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
