% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_Definitions.R
\docType{data}
\name{TableDef_methods}
\alias{TableDef_methods}
\alias{DataManR}
\title{TableDef Methods}
\format{
An object of class \code{proto} (inherits from \code{environment}) of length 25.
}
\usage{
DataManR
}
\arguments{
\item{data}{data to write. Defaults to data field.}

\item{file}{file to write. Defaults to file field.}

\item{sep}{delimiter to write as. Defaults to ",".}

\item{append}{logical value. Defaults to FALSE.}
}
\description{
Methods of TableDef
}
\keyword{datasets}
